requirements.txt
asgiref==3.8.1
black==25.9.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
cryptography==46.0.3
Django==5.0.7
django-allauth==65.13.0
django-braces==1.15.0
idna==3.8
PyJWT>=2.8.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pycparser==2.23
python-decouple==3.8
pytokens==0.1.10
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.2.2
